requirements.txt
streamlit
pandas
mecab-python3
numpy
plotly
google-generativeai